+++ src/main/resources/templates/account.mustache
@@ -23,7 +23,7 @@
 
         <ul>
             <div class="LOGO">НАШ КИНО ТЕАТР</div>
-            <a href="index.html" class="buttonAFISHA">КИНОАФИША</a>
+            <a href="index" class="buttonAFISHA">КИНОАФИША</a>
             <a href="#" class="buttonCINEMAS">КИНОТЕАТРЫ</a>
         </ul>
     </nav>
@@ -34,7 +34,7 @@
     <a href="#" id="tab_name">ПРОФИЛЬ</a>
     <div class="row">
         <div class="col-6 btn_group_menu">
-            <a href="index.html" class="buttonAFISHA">КИНОАФИША</a>
+            <a href="index" class="buttonAFISHA">КИНОАФИША</a>
             <a href="#" class="buttonCINEMAS">КИНОТЕАТРЫ</a>
         </div>
     </div>
